+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 ## 0.1.1-SNAPSHOT
 
-- ?
+- #1 Load printing namespaces for EDN
+- #2 add more types: Agent Volatile Namespace
 
 ## 0.1.0